+++ server/apps/core/src/modules/inventory/inventory.module.ts
@@ -1,5 +1,3 @@
-import { ApplicationConfig } from '@configuration/environtment'
-import { MongoConfig } from '@configuration/mongo'
 import { AccountModule } from '@core/account/account.module'
 import { Account, AccountSchema } from '@core/account/schemas/account.model'
 import { GeneralReceiveNoteController } from '@core/inventory/general.receive.note.controller'
@@ -7,22 +5,6 @@ import { GeneralReceiveNoteService } from '@core/inventory/general.receive.note.
 import { InventoryService } from '@core/inventory/inventory.service'
 import { PurchaseOrderController } from '@core/inventory/purchase.order.controller'
 import { PurchaseOrderService } from '@core/inventory/purchase.order.service'
-import {
-  GeneralReceiveNote,
-  GeneralReceiveNoteSchema,
-} from '@core/inventory/schemas/general.receive.note'
-import {
-  PurchaseOrder,
-  PurchaseOrderSchema,
-} from '@core/inventory/schemas/purchase.order'
-import {
-  InventoryStock,
-  InventoryStockSchema,
-} from '@core/inventory/schemas/stock'
-import {
-  InventoryStockLog,
-  InventoryStockLogSchema,
-} from '@core/inventory/schemas/stock.log'
 import { MasterItemService } from '@core/master/master.item.service'
 import { MasterStockPointService } from '@core/master/master.stock.point.service'
 import { MasterItem, MasterItemSchema } from '@core/master/schemas/master.item'
@@ -34,152 +16,30 @@ import {
   MasterStockPoint,
   MasterStockPointSchema,
 } from '@core/master/schemas/master.stock.point'
+import {
+  GeneralReceiveNote,
+  GeneralReceiveNoteSchema,
+} from '@inventory/schemas/general.receive.note'
+import {
+  PurchaseOrder,
+  PurchaseOrderSchema,
+} from '@inventory/schemas/purchase.order'
+import { InventoryStock, InventoryStockSchema } from '@inventory/schemas/stock'
+import {
+  InventoryStockLog,
+  InventoryStockLogSchema,
+} from '@inventory/schemas/stock.log'
 import { LogActivity, LogActivitySchema } from '@log/schemas/log.activity'
 import { LogLogin, LogLoginSchema } from '@log/schemas/log.login'
 import { Module } from '@nestjs/common'
-import { ConfigModule } from '@nestjs/config'
+import { ClientsModule } from '@nestjs/microservices'
 import { MongooseModule } from '@nestjs/mongoose'
 import { AuthModule } from '@security/auth.module'
-import { TimeManagement } from '@utility/time'
+import { KafkaConn } from '@utility/kafka'
 
 @Module({
   imports: [
-    ConfigModule.forRoot({
-      isGlobal: true,
-      envFilePath: `${process.cwd()}/environment/${
-        process.env.NODE_ENV === 'development' ? '' : process.env.NODE_ENV
-      }.env`,
-      load: [ApplicationConfig, MongoConfig],
-    }),
-    MongooseModule.forFeatureAsync([
-      {
-        name: InventoryStock.name,
-        useFactory: () => {
-          const schema = InventoryStockSchema
-          const time = new TimeManagement()
-          schema.pre('save', function (next) {
-            if (this.isNew) {
-              this.__v = 0
-            }
-
-            if (this.isModified()) {
-              this.increment()
-              return next()
-            } else {
-              return next(new Error('Invalid document'))
-            }
-          })
-
-          schema.pre('findOneAndUpdate', async function (next) {
-            const update = this.getUpdate()
-            const docToUpdate = await this.model.findOne(this.getQuery())
-            if (docToUpdate) {
-              update['updated_at'] = time.getTimezone('Asia/Jakarta')
-              update['$inc'] = { __v: 1 }
-            }
-            next()
-          })
-
-          return schema
-        },
-      },
-      {
-        name: PurchaseOrder.name,
-        useFactory: () => {
-          const schema = PurchaseOrderSchema
-          const time = new TimeManagement()
-          schema.pre('save', function (next) {
-            if (this.isNew) {
-              this.id = `purchase_order-${this._id}`
-              this.__v = 0
-            }
-
-            if (this.isModified()) {
-              this.increment()
-              this.updated_at = time.getTimezone('Asia/Jakarta')
-              return next()
-            } else {
-              return next(new Error('Invalid document'))
-            }
-          })
-
-          schema.pre('findOneAndUpdate', function (next) {
-            const update = this.getUpdate()
-            update['updated_at'] = time.getTimezone('Asia/Jakarta')
-            update['$inc'] = { __v: 1 }
-            next()
-          })
-
-          return schema
-        },
-      },
-      {
-        name: GeneralReceiveNote.name,
-        useFactory: () => {
-          const schema = GeneralReceiveNoteSchema
-          const time = new TimeManagement()
-          schema.pre('save', function (next) {
-            if (this.isNew) {
-              this.id = `general_receive_note-${this._id}`
-              this.__v = 0
-            }
-
-            if (this.isModified()) {
-              this.increment()
-              this.updated_at = time.getTimezone('Asia/Jakarta')
-              return next()
-            } else {
-              return next(new Error('Invalid document'))
-            }
-          })
-
-          schema.pre('findOneAndUpdate', function (next) {
-            const update = this.getUpdate()
-            update['updated_at'] = time.getTimezone('Asia/Jakarta')
-            update['$inc'] = { __v: 1 }
-            next()
-          })
-
-          return schema
-        },
-      },
-      {
-        name: MasterItemBatch.name,
-        useFactory: () => {
-          const schema = MasterItemBatchSchema
-          const time = new TimeManagement()
-          schema.pre('save', function (next) {
-            if (this.isNew) {
-              this.id = `item_batch-${this._id}`
-              this.__v = 0
-            }
-
-            if (this.isModified()) {
-              this.increment()
-              this.updated_at = time.getTimezone('Asia/Jakarta')
-              return next()
-            } else {
-              return next(new Error('Invalid document'))
-            }
-          })
-
-          schema.pre('findOneAndUpdate', async function (next) {
-            const update = this.getUpdate()
-            const docToUpdate = await this.model.findOne(this.getQuery())
-            if (docToUpdate) {
-              update['updated_at'] = time.getTimezone('Asia/Jakarta')
-              update['$inc'] = { __v: 1 }
-            } else {
-              update['id'] = `item_batch-${update['_id']}`
-            }
-
-            next()
-          })
-
-          return schema
-        },
-      },
-    ]),
+    ClientsModule.registerAsync([KafkaConn.inventory[0]]),
     MongooseModule.forFeature([
       { name: Account.name, schema: AccountSchema },
       { name: LogLogin.name, schema: LogLoginSchema },
@@ -187,6 +47,9 @@ import { TimeManagement } from '@utility/time'
       { name: MasterItem.name, schema: MasterItemSchema },
       { name: MasterItemBatch.name, schema: MasterItemBatchSchema },
       { name: MasterStockPoint.name, schema: MasterStockPointSchema },
+      { name: PurchaseOrder.name, schema: PurchaseOrderSchema },
+      { name: GeneralReceiveNote.name, schema: GeneralReceiveNoteSchema },
+      { name: InventoryStock.name, schema: InventoryStockSchema },
       { name: InventoryStockLog.name, schema: InventoryStockLogSchema },
     ]),
     AuthModule,
