+++ server/apps/inventory/src/dto/stock.ts
