+++ server/apps/inventory/src/schemas/general.receive.note.detail.ts
