+++ server/apps/inventory/src/schemas/purchase.order.detail.ts
