+++ server/apps/inventory/src/schemas/stock.log.ts
