+++ server/apps/inventory/src/schemas/stock.ts
