+++ server/apps/utility/dto/file.ts
@@ -1,0 +1,6 @@
+export interface FileDto {
+  filename: string
+  path: string
+  mimetype: string
+  buffer: Buffer
+}
