+++ server/package.json
@@ -10,6 +10,7 @@
     "format": "prettier --write \"apps/**/*.ts\" \"libs/**/*.ts\"",
     "start": "nest start",
     "start:dev:core": "nest build core --webpack --webpackPath ./webpack.config.js --watch",
+    "start:dev:inventory": "nest build inventory --webpack --webpackPath ./webpack.config.js --watch",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/apps/server/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
@@ -55,10 +56,13 @@
     "@typescript-eslint/parser": "^5.53.0",
     "bcrypt": "^5.1.0",
     "copy-webpack-plugin": "^11.0.0",
+    "dotenv": "^16.0.3",
     "eslint": "^8.34.0",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-prettier": "^4.2.1",
+    "fs": "^0.0.1-security",
     "jest": "29.4.3",
+    "kafkajs": "^2.2.4",
     "mongodb-memory-server": "^8.11.5",
     "passport": "^0.6.0",
     "prettier": "^2.8.4",
@@ -84,6 +88,7 @@
     ],
     "moduleNameMapper": {
       "^@core(.*)$": "<rootDir>/apps/core/src/modules$1",
+      "^@inventory(.*)$": "<rootDir>/apps/inventory/src$1",
       "^@log(.*)$": "<rootDir>/apps/log/src$1",
       "^@security(.*)$": "<rootDir>/apps/security$1",
       "^@configuration(.*)$": "<rootDir>/apps/configuration$1",
